requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
gunicorn>=20.1.0
pydantic>=1.10.7
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
# Remove PyJWT if it's causing conflicts
# jwt>=1.3.1  

# Authentication/Authorization
passlib[bcrypt]>=1.7.4

# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.9
asyncpg>=0.27.0  # Async PostgreSQL driver

# Email and validation
email-validator>=2.0.0
jinja2>=3.1.2

# WebSockets
websockets>=11.0.2

# Utilities
python-dotenv>=1.0.0
user-agents>=2.2.0
aiocache>=0.12.1
